Use mongodb-stitch-server-sdk for build time calls
