+++ src/utils/fetch-youtube-data.js
@@ -1,27 +1,12 @@
-import dlv from 'dlv';
 import { requestYoutubePlaylist } from './devhub-api-stitch';
+import { simplifyYoutubeResponse } from './simplify-youtube-response';
 // Fetches data from youtube api
-
-const simplifyResponse = responseData => {
-    const video = responseData.snippet;
-    const youtubeJSON = {
-        mediaType: 'youtube',
-        title: video['title'],
-        publishDate: video['publishedAt'],
-        description: video['description'],
-        videoId: dlv(video, 'resourceId.videoId', []),
-        playlistId: video['playlistId'],
-        thumbnailUrl: dlv(video, 'thumbnails.standard.url', []),
-    };
-
-    return youtubeJSON;
-};
 
 const fetchYoutubeData = async (maxResults = 5) => {
     try {
         const response = await requestYoutubePlaylist(maxResults);
         if (response) {
-            const videoList = response.items.map(simplifyResponse);
+            const videoList = response.items.map(simplifyYoutubeResponse);
             return videoList;
         }
     } catch (e) {
