+++ src/utils/setup/handle-create-page.js
@@ -3,6 +3,7 @@ import { removeExcludedArticles } from './remove-excluded-articles';
 import { removePageIfStaged } from './remove-page-if-staged';
 import { getNestedValue } from '../get-nested-value';
 import { getMetadata } from '../get-metadata';
+import { fetchBuildTimeVideos } from './fetch-build-time-videos';
 
 const metadata = getMetadata();
 let stitchClient;
@@ -145,12 +146,14 @@ export const handleCreatePage = async (
                 learnFeaturedArticles || DEFAULT_FEATURED_LEARN_SLUGS,
                 MAX_LEARN_PAGE_FEATURED_ARTICLES
             );
+            const allVideos = await fetchBuildTimeVideos();
             deletePage(page);
             createPage({
                 ...page,
                 context: {
                     ...page.context,
                     allArticles: learnPageArticles,
+                    allVideos,
                     featuredArticles: featuredLearnArticles,
                     filters,
                 },
