+++ src/utils/setup/init-stich.js
@@ -1,10 +1,10 @@
 import { Stitch, AnonymousCredential } from 'mongodb-stitch-server-sdk';
 import { SNOOTY_STITCH_ID } from '../../build-constants';
 
-export const initStitch = async () => {
-    const stitchClient = Stitch.hasAppClient(SNOOTY_STITCH_ID)
-        ? Stitch.getAppClient(SNOOTY_STITCH_ID)
-        : Stitch.initializeAppClient(SNOOTY_STITCH_ID);
+export const initStitch = async (id = SNOOTY_STITCH_ID) => {
+    const stitchClient = Stitch.hasAppClient(id)
+        ? Stitch.getAppClient(id)
+        : Stitch.initializeAppClient(id);
     await stitchClient.auth
         .loginWithCredential(new AnonymousCredential())
         .catch(console.error);
